+++ agent/src/main/java/com/teamscale/jacoco/agent/upload/teamscale/TeamscaleConfig.java
@@ -12,8 +12,9 @@ import com.teamscale.report.util.BashFileSkippingInputStream;
 import com.teamscale.report.util.ILogger;
 
 import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
+import java.nio.file.Files;
+import java.util.List;
 import java.util.jar.JarInputStream;
 import java.util.jar.Manifest;
 
@@ -147,7 +148,7 @@ public class TeamscaleConfig {
 	 */
 	private Manifest getManifestFromJarFile(File jarFile) throws AgentOptionParseException {
 		try (JarInputStream jarStream = new JarInputStream(
-				new BashFileSkippingInputStream(new FileInputStream(jarFile)))) {
+				new BashFileSkippingInputStream(Files.newInputStream(jarFile.toPath())))) {
 			Manifest manifest = jarStream.getManifest();
 			if (manifest == null) {
 				throw new AgentOptionParseException(
@@ -164,14 +165,20 @@ public class TeamscaleConfig {
 												String value) throws AgentOptionParseException {
 		File jarFile = filePatternResolver.parsePath(optionName, value).toFile();
 		try {
-			String revision = GitPropertiesLocatorUtils.getRevisionFromGitProperties(jarFile, true);
-			if (revision == null) {
-				throw new AgentOptionParseException("Could not locate a git.properties file in " + jarFile.toString());
+			// We can't be sure that the search-git-properties-recursively option is parsed already.
+			// Since we only support one git.properties file here anyway, recursive search is disabled.
+			List<String> revisions = GitPropertiesLocatorUtils.getRevisionsFromGitProperties(jarFile, true, false);
+			if (revisions.isEmpty()) {
+				throw new AgentOptionParseException("Found no git.properties files in " + jarFile);
 			}
-			return revision;
+			if (revisions.size() > 1) {
+				throw new AgentOptionParseException("Found multiple git.properties files in " + jarFile +
+						". Uploading to multiple projects is currently not possible with the option teamscale-git-properties-jar. " +
+						"Please contact CQSE if you need this feature.");
+			}
+			return revisions.get(0);
 		} catch (IOException | InvalidGitPropertiesException e) {
-			throw new AgentOptionParseException("Could not locate a valid git.properties file in " + jarFile.toString(),
-					e);
+			throw new AgentOptionParseException("Could not locate a valid git.properties file in " + jarFile, e);
 		}
 	}
 }